Fix fees in admin flow for pairing transactions (#353)
* Fix fees in admin flow for pairing transactions

Thus far, when admin saves a transaction to pair it, Bank would
automatically assume any positive transactions must be charged a Bank
fee, which is often not the case. This led to two confusing cases: one,
where the admin expects no fees based on the unchecked field in the form
but finds that a fee was charged, and two, where the admin wants to
ensure that no fees are charged, but must first pair the TX before
uchecking the field and re-saving.

This commit removes the old behavior of forcibly checking the
fee-applies checkbox when TXs are paired, and instead pre-fills the form
when a TX is first paired, as a suggestion to the admin. The admin may
then choose to uncheck it.

* Only prefill charge fees checkbox if a TX is event-related