Dupe icon svg from cache before making possible mutations (#2497)
Fixes https://github.com/hackclub/bank/issues/2494

Bug introduced in https://github.com/hackclub/bank/pull/2482. I didn't
realize we mutate the parsed nokogiri object that's stored in the cache
so we could end up having incorrect styles. This fixes it by creating a
dupe of the parsed object before mutating for styles.

This doesn't seem to have affected endpoint time that much
```
  Rendered collection of canonical_transactions/_canonical_transaction.html.erb [100 times] (Duration: 43.8ms | Allocations: 41525)
```

This line is roughly the same before and after this change, which is
still an improvement over the ~200ms we were seeing without any caching