Remove use of lets and explicitly create test data per test
As I mentioned in the previous commit message, lets can be considered an
antipattern so here is the alternative, creating the test data
explicitly.

I created several helper methods for doing this here, but we can replace
these with factories.

Although this change does look like it's violating DRY, I recommend
against the urge to DRY up test code as it hurts maintainability in the
long run.