Remove `#` from G Suite Organizational Unit Name (#2663)
Seems like having multiple `#` in the OU Name is an issue because the Google Ruby SDK doesn't escape the `#` character. This causes an `URI::InvalidURIError:  bad URI(is not URI?)` error when the SDK forms a URL for an API request.
The old URLs looks something like this: `https://admin.googleapis.com/admin/directory/v1/customer/<REDACTED>/orgunits/Events/?#1558%20G#1110`
This old URL format was working for a period of time, however, I think upgrading our Ruby version included a change to the URI parser which broke this!