Add annotate gem to improve developer workflow (#2662)
Closes #2656

Install https://github.com/ctran/annotate_models gem, which automatically comments model files with the database fields. In my experience this is nicer for developers than having to check schema.rb or the database directly, so I recommend we try it out and see if that's the case.

These annotate settings will update the files whenever migrations are run, so they should stay in sync with the db schema.