Introduce factory_bot for creating test data instead of fixtures (#2660)
This PR introduces `factory_bot` as an alternative to creating test data with fixtures.
See https://semaphoreci.com/blog/2014/01/14/rails-testing-antipatterns-fixtures-and-factories.html for more opinionated thoughts on using fixtures (from the linked article: `Do not use fixtures.` 😆 )

I made this as a separate PR from https://github.com/hackclub/bank/pull/2657 to explicitly show the difference between creating test data with the model vs factories. IMO the biggest benefit is that you can rely on some sane defaults to wire up associations automatically as well as have some setup that is required to make the object valid (but is not relevant to the test). The negatives are that the factory_bot DSL can be pretty weird and it's sometimes difficult to go from a factory creation to where it's defined, and also if creating a set of data in a certain state is managed by a service object in the application and that service object is already well tested, I would recommend using that instead to setup the test data instead of trying to duplicate that complexity in factory definitions

* Install factory_bot and faker gems

Per recommendation on factory_bot docs, use factory_bot_rails and
set up a spec/support/factory_bot.rb
https://github.com/thoughtbot/factory_bot/blob/master/GETTING_STARTED.md#rspec

* Create factory for partner, event and disbursement