Fix and refactor receipt_uploads_mailbox_spec (#2724)
* Use event factory instead of fixture

This is sufficient to fix #2703 because we are creating a new event each
time and so we don't run into this validation error

```
Failure/Error: event.users << user

ActiveRecord::RecordInvalid:
 Validation failed: User has already been taken
```

* Create minimal canonical_transaction factory

* Create minimal canonical_event_mapping factory

* Refactor receipt_uploads_mailbox_spec to use factories

As with other tests with this refactoring, I kept the `let` structure as
is for now, except this was a good example of how the laziness of `let`
can burn you. Because `let` is lazy, `canonical_event_mapping` would not
get created in one of the specs and thus the user would not have the
correct permissions on the event. For now I just changed everything to
`let!` to eagerly create everything

* Upgrade rspec-rails to 5 and uncomment text broken by older version

According to the rspec-rails README, we should use
`rspec-rails 5 supports Rails 5.2 to 6.1.`