Add presence validation for event.organization_identifier (#2725)
I noticed this while adding the factory for events. We have a database
constraint on this column already but application validators tend to
have friendlier error messages. Unfortunately ActiveRecord does not have
a good way to avoid this duplication, so I would argue we try to add
missing validations everywhere or if are only going to pick one, the
database constraints are more valuable.

Some reading on this topic:
- https://thoughtbot.com/blog/validation-database-constraint-or-both
- https://naildrivin5.com/blog/2015/11/15/rails-validations-vs-postgres-check-constraints.html
- https://boringrails.com/tips/rails-check-constraints-database-validations
- https://bignerdranch.com/blog/dont-get-tangled-in-your-database-constraints/