Unskip UserService::Create spec and migrate to factories (#3010)
https://github.com/hackclub/bank/issues/3008

* Unskip UserService::Create

Test passes after skipping call to BankApi service

* Get tests to pass with factories instead of fixtures

This already revealed some hidden assumptions from the fixtures, such
as that user1 was an organizer in event1.

Also, let! was required instead of let because the laziness could cause
the changed count for User and OrganizerPosition to be overcounted since
it would include users or organizer_positions from the original factory creation.
This is one of the downsides of heavy use of lets.

* Clean up lets usage

No more let! required to get the test to pass
contexts share less setup (only event creation)

Also added a missing set of tests, for the case where a user already
exists but is not already an organizer on the event