Use --data-only pg_dump for creating review apps (#3311)
This is an attempt to fix https://github.com/hackclub/bank/issues/3205 based on https://help.heroku.com/ZOFBHJCJ/heroku-postgres-extension-changes-faq

As mentioned in the issue, the problem centers around installing extensions. Something I noticed during my investigation is that `bin/rails db:create` was succeeding (which would create the `g_suites` table), but then the `pg_dump`/`pg_restore` would try to drop and recreate everything. So based on this, I wondered if we could do a data only dump/restore from postgres so that we use the schema from the rake task and just load data into it. Turns out there is a `--data-only` flag for `pg_dump`.

This seems to produce a working review app!