Add Card Charges to the API (#3349)
Card Charges are not an actual linked object within the transaction engine. It's pretty messy. In the TX engine, they're represented with:
- StripeAuthorization
- RawPendingStripeTransaction
- RawStripeTransaction

and accessed via HcbCode using:
- HashedTransaction
- CanonicalTransaction
- CanonicalPendingTransaction

There isn't a great Model for representing Card Charges. However, since HcbCodes have easy access to all the card charge data, we'll be using HcbCodes to represent Card Charges in the API.

This PR adds an Api::Models::CardCharge class that is an ActiveRecord model and inherits from HcbCode. This definitely feels pretty hacky, but it seems to be the simplest way to give card charges their own public ID and to have "stricter" types. I've tried my best to prevent database modifications by this CardCharge model.