Fix enum Ruby 3 keyword argument warnings (#3898)
`check_deposit` and `recurring_donation` are false positives, because we
are getting a complaint that the hash representing the enum values is trying
to call the method with keyword arguments (which doesn't make sense
because they are developer provided values, not keyword arguments
specified by the enum method). To silence these errors for the time
being, we can add `suffix: false` as the last option which is already
the default value, but means the last argument is not a hash.

For `country_enumable`, it looks like bank is metaprogramming with the
old rails syntax using a key value pair for enum name and values instead
of separate arguments. See
https://blog.saeloun.com/2021/02/26/rails-introduces-new-syntax-for-enum/
So I updated that to the modern syntax, and also added `suffix: false`
for the same reasons as above.

We can verify these settings are producing the same enum configuration
as well as helper methods by checking this in console before and after

```
RecurringDonation.stripe_statuses
Event.countries
CheckDeposit.increase_statuses

RecurringDonation.methods - Object.methods
Event.methods - Object.methods
CheckDeposit.methods - Object.methods
```