Show Images / PDFs in Transaction Popover (#4175)
Instead of creating a second modal, in this PR, after you click a
preview of a receipt it will hide the other elements in the transaction
popover and make the receipt larger. If the receipt is an image, it'll
just be larger. If the receipt is a PDF, it'll be a full embedded PDF.
The size are both PDFs or images are roughly equivalent to what their
size would have been if they were in a second modal. There's also a
minimize receipt button shown during expansion.

<img width="766" alt="Screenshot 2023-08-21 at 12 05 42 PM"
src="https://github.com/hackclub/bank/assets/39828164/f7b12bb2-7118-4cdc-adb6-571be2369e3b">

<img width="831" alt="Screenshot 2023-08-21 at 12 06 24 PM"
src="https://github.com/hackclub/bank/assets/39828164/687651ea-3f57-4434-a351-7d5555d5e274">

Fixes #4140

---------

Co-authored-by: yodalightsabr <yoda@yodacode.xyz>