Reply to receipt upload emails (#4786)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

When you email HCB with a receipt, the confirmation back from HCB shows
up as a separate thread rather than a reply.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

This PR sets the `In-Reply-To` header, which should theoretically tell
mail clients to treat these messages as replies. I've tested this and
verified that the header is set correctly.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

Resolves #4769