Install `rubocop-rails`! (#4789)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

linters are exciting! i wish we had more linters.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

This adds the `rubocop-rails` plugin, which introduces a bunch of rules
that identify Rails specific problems & deprecation. Not all of these
rules are helpful, but many are useful for fixing common performance and
consistency issues.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->