Fix uploaded profile picture deletion (#4855)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Makes profile picture deletion work.
Fixes #4854 


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
When I clicked the link to delete the profile picture in development
mode, the specific error was that it could not find a user (at the
beginning of the `delete_profile_picture` function in
`users_controller.rb`. It also showed the params of the function when it
was called, which included the user id as `user_id` instead of `id`. I'm
not sure why the key is different for this specific function in
`users_controller.rb`, but it seemed to fix it when I changed it.