Admin: User information page (#4798)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
This completes Gary's request in #4651 by adding a more complete page to
see any relevant information about a user


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
I decided to use the admin controller for this page as I concluded it's
an admin page, so it's protected by that authentication level, It also
appears to take an id parameter for all its routes too.
I then searched each model using the user ID. and put that information
in a table so that it only shows if the user has 1 of those
transactions/cards

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

![image](https://github.com/hackclub/hcb/assets/81807361/617c1879-8841-4301-a28a-6b4adc2d19b8)

---------

Co-authored-by: Sarthak Mohanty <sarthaktexas@gmail.com>
Co-authored-by: Gary Tou <gary@garytou.com>