Email commands: now we have more! (#5196)
## Summary of the problem

Email commands are cool but we only have one. Now we have `@missing` and
`@tag` as well.

We also had a couple of bugs, as shown by this HCB code:
https://hcb.hackclub.com/hcb/GkHWr4g.

Closes https://github.com/hackclub/hcb/issues/5336

## Describe your changes

<img width="759" alt="Screenshot 2024-02-04 at 11 45 25 AM"
src="https://github.com/hackclub/hcb/assets/39828164/80d05a32-fe81-4656-89f8-69728250fb83">

We also handle the bug where it would upload the email with the commands
in it as a file. And where we'd just rename the transaction to "to
rename the transaction."

---------

Co-authored-by: Gary Tou <gary@garytou.com>