Refactor check association to use belongs_to instead of find_by
Given that the type of the foreign key (check_transaction_id) is
a string (varchar) and not a integer (bigint) like the id of the
check table and the fact that there is no foreign key constraint
between the two tables, I'm not sure if the usage of `find_by`
instead of an ActiveRecord association is intentional or not.

Let me know if I'm missing something here.