Add a "Search" action for the Command Bar (#4839)
## Summary of the problem

Wouldn't it be incredible if we could use the Command Bar to "search"
HCB? Imagine searching for a card or a transaction or a donation or...
anything on HCB!

## Describe your changes

This is a draft pull request that models out how this can be achieved
using the `kbar` package. Here's a visual demo of how it works:


https://github.com/hackclub/hcb/assets/39828164/a9ad535c-8caf-468b-ac77-0999e50045be

To make this shippable, we would need to:

* Expand the SearchController to... well search HCB! At the moment it's
just a dummy API.
* Would using Caleb's API make sense here instead of implementing a new
search API?
* Map out how users select "what" they're searching for, how do we
distinguish between searching for cards, donations, or transactions?
Right now we... can't.
* One option would be having people add modifiers to their search, eg.
`@card`:
<img width="537" alt="Screenshot 2023-12-23 at 12 17 06 AM"
src="https://github.com/hackclub/hcb/assets/39828164/3d0e4c35-a0a4-455a-b329-8fdc9dafe88f">
If we were to do this, we could add a bunch of additional modifiers! So
many modifiers... we would just need to design a custom search input for
`kbar` that adds fancy styles to these modifiers.
    
* Another, less flexible but easier to implement solution, would be to
have people select what they'd want to search before searching. If
possible, I'd prefer to have a modifiers setup.
<img width="745" alt="Screenshot 2023-12-23 at 12 22 39 AM"
src="https://github.com/hackclub/hcb/assets/39828164/2d310666-2e29-4a91-a25e-6220a55ccd33">

* We'd need to make some custom styles for this, espc. if we got with
the modifiers setup.

But yeah! Here's a potential structure for this sort of thing... was a
funny little problem to try and figure out.

There's probably a world where we could replace the whole command bar
with search... this doesn't quite do that but it lays the foundations.

---------

Co-authored-by: yodalightsabr <yoda@yodacode.xyz>
Co-authored-by: Gary Tou <gary@garytou.com>