[Sidekiq] Fix redirect to auth when non-admin (#8038)
It broke after moving auth to the LoginsController. It also previously
didn't redirect; when we should.


Also, I've refactored the AdminConstraint usage into a block. Previously
it was added as an argument to multiple `mount`s. Now, the block will
apply the constraint to everything inside it.

To make the code cleaner, I've also made `AdminConstraint#matches?` a
class method to prevent having to call `new` in the routes file.