[IncreaseCheckMailer] Set reply_to for notify_recipient and remind_recipient (#8355) (#8420)
### See [`#8355`](https://github.com/hackclub/hcb/issues/8355)

This pull request includes updates to the `IncreaseCheckMailer` class to
enhance the email functionalities by adding a `reply_to` field.

Enhancements to email functionalities:

*
[`app/mailers/increase_check_mailer.rb`](diffhunk://#diff-399bdfa7e2f4380cb926af7b71650f42a1529605c4b36ea60a73b4056b785340L7-R19):
Added a `reply_to` field in the `notify_recipient` and
`remind_recipient` methods to include the email addresses of event
organizers with the role of manager.

---------

Co-authored-by: Gary Tou <gary@garytou.com>