[Disbursements] Improve organization selector (#8136)
## Summary of the problem

It can be difficult to select organizations as an admin (especially on
mobile) because the native dropdown lists thousands of organizations.


## Describe your changes

I've built a custom dropdown component we can use on the disbursements
page (and in the future, elsewhere, too).

For admins, it shows the organization ID and moves the balance into a
tooltip.

Here's a video:


https://github.com/user-attachments/assets/0e893cc8-62db-462b-be5e-9ae4086c60f1

And here's a screenshot on dark mode:

<img width="479" alt="image"
src="https://github.com/user-attachments/assets/96229c44-6bde-4482-beb1-353dab219c72">